# src/main/resources/public/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Service REST API Spec",
    "version" : "\"v1\"",
    "title" : "Data Mining Competition"
  },
  "host" : "localhost:8081",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/change-password" : {
      "put" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Change player password",
        "description" : "",
        "operationId" : "changePassForLoggedUser",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PasswordRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Validation error, more information will be provided in sent JSON"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    },
...
    "/competitions/created" : {
      "get" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Get all created competitions name",
        "description" : "",
        "operationId" : "getAllCreatedCompetitions",
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/CompetitionCollectionDto"
              }
            }
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    },
    "/competitions/participation" : {
      "get" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Get all competitions that user participate",
        "description" : "",
        "operationId" : "getAllParticipation",
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ParticipationDto"
              }
            }
          },
          "400" : {
            "description" : "Validation error, more information will be provided in sent JSON"
          },
          "404" : {
            "description" : "User not found"
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    },
...
    "/profile" : {
      "get" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Get profile",
        "description" : "",
        "operationId" : "getProfile",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UserDto"
            }
          },
          "500" : {
            "description" : "Something went wrong in Server"
          }
        }
      },
      "put" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Update profile",
        "description" : "",
        "operationId" : "updateProfile",
        "produces" : [ "application/json", "text/plain" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object"
            }
          },
          "400" : {
            "description" : "Validation error, more information will be provided in sent json"
          },
          "500" : {
            "description" : "Something went wrong in Server"
          }
        }
      }
    },
...
    "/users" : {
      "get" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Get all users. You will need to have admin scope",
        "description" : "",
        "operationId" : "getAllUsers",
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ManageUserDto"
              }
            }
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    },
    "/users/{id}" : {
      "delete" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Delete user by id. You will need to have admin scope",
        "description" : "",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    },
    "/users/{id}/active" : {
      "put" : {
        "tags" : [ "user : Operations about user" ],
        "summary" : "Update isActive state for user. You will need to have admin scope",
        "description" : "",
        "operationId" : "setActive",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ManageActiveDto"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Something wrong in Server"
          }
        }
      }
    }
  },
  "securityDefinitions" : {
    "api_key" : {
      "type" : "apiKey"
    }
  },
  "definitions" : {
    "CompetitionCollectionDto" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        },
        "shortDescription" : {
          "type" : "string"
        },
        "author" : {
          "type" : "string"
        }
      }
    },
...
    "ManageActiveDto" : {
      "type" : "object",
      "properties" : {
        "isActive" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
    "PasswordRequest" : {
      "type" : "object",
      "properties" : {
        "oldPassword" : {
          "type" : "string"
        },
        "newPassword" : {
          "type" : "string"
        }
      }
    },
...
    "UserDto" : {
      "type" : "object",
      "properties" : {
        "username" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 50
        },
        "firstName" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 50
        },
        "lastName" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 50
        },
        "email" : {
          "type" : "string",
          "minLength" : 5,
          "maxLength" : 100
        },
        "city" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 50
        },
        "country" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 50
        },
        "postalCode" : {
          "type" : "string",
          "minLength" : 0,
          "maxLength" : 10
        },
        "password" : {
          "type" : "string",
          "minLength" : 4,
          "maxLength" : 100
        },
        "activated" : {
          "type" : "boolean",
          "default" : false
        },
        "authorities" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    }
  }
}